cmd/govim: refactor functionality of govim into separate files
Just makes working with it easier.
